Modulify AttendanceDetails and add smoke test describing empty data conditions
