+++ ui/legacy.js
@@ -29,7 +29,6 @@ import '../app/assets/javascripts/student_profile/pdf/student_profile_pdf.js';
   // details:
 import '../app/assets/javascripts/student_profile/services_details.jsx';
 import '../app/assets/javascripts/student_profile/profile_details.jsx';
-import '../app/assets/javascripts/student_profile/attendance_details.jsx';
 import '../app/assets/javascripts/student_profile/ela_details.jsx';
 import '../app/assets/javascripts/student_profile/math_details.jsx';
